+++ containeranalysis/google/cloud/devtools/containeranalysis_v1/gapic/container_analysis_client_config.py
@@ -1,0 +1,38 @@
+config = {
+    "interfaces": {
+        "google.devtools.containeranalysis.v1.ContainerAnalysis": {
+            "retry_codes": {
+                "idempotent": ["DEADLINE_EXCEEDED", "UNAVAILABLE"],
+                "non_idempotent": [],
+            },
+            "retry_params": {
+                "default": {
+                    "initial_retry_delay_millis": 100,
+                    "retry_delay_multiplier": 1.3,
+                    "max_retry_delay_millis": 60000,
+                    "initial_rpc_timeout_millis": 20000,
+                    "rpc_timeout_multiplier": 1.0,
+                    "max_rpc_timeout_millis": 20000,
+                    "total_timeout_millis": 600000,
+                }
+            },
+            "methods": {
+                "SetIamPolicy": {
+                    "timeout_millis": 30000,
+                    "retry_codes_name": "non_idempotent",
+                    "retry_params_name": "default",
+                },
+                "GetIamPolicy": {
+                    "timeout_millis": 30000,
+                    "retry_codes_name": "non_idempotent",
+                    "retry_params_name": "default",
+                },
+                "TestIamPermissions": {
+                    "timeout_millis": 30000,
+                    "retry_codes_name": "non_idempotent",
+                    "retry_params_name": "default",
+                },
+            },
+        }
+    }
+}