+++ datacatalog/CHANGELOG.md
@@ -4,6 +4,27 @@
 
 [1]: https://pypi.org/project/google-cloud-datacatalog/#history
 
+## 0.2.0
+
+06-12-2019 12:46 PDT
+
+### New Features
+
+- Add search capability, tags that match a query, and IAM policies ([#8266](https://github.com/googleapis/google-cloud-python/pull/8266))
+- Add protos as an artifact to library (via synth). ([#8018](https://github.com/googleapis/google-cloud-python/pull/8018))
+
+### Documentation
+
+- Add nox session `docs`, reorder methods (via synth). ([#7766](https://github.com/googleapis/google-cloud-python/pull/7766))
+- Fix broken link to client library docs in README ([#7713](https://github.com/googleapis/google-cloud-python/pull/7713))
+
+### Internal / Testing Changes
+
+- Suppress checking 'cov-fail-under' in nox default session (via synth). ([#8235](https://github.com/googleapis/google-cloud-python/pull/8235))
+- Fix coverage in 'types.py' (via synth). ([#8150](https://github.com/googleapis/google-cloud-python/pull/8150))
+- Blacken noxfile.py, setup.py (via synth). ([#8117](https://github.com/googleapis/google-cloud-python/pull/8117))
+- Add empty lines (via synth). ([#8052](https://github.com/googleapis/google-cloud-python/pull/8052))
+
 ## 0.1.0
 
 04-15-2019 15:46 PDT